+++ src/main/java/com/gmail/kramarenko104/repository/UserCrudRepository.java
@@ -6,15 +6,15 @@ import java.util.List;
 
 public interface UserCrudRepository {
 
-    int createUser(User user);
+    long createUser(User user);
 
-    User getUser(int id);
+    User getUser(long id);
 
     User getUserByLogin(String login);
 
-    int updateUser(User user);
+    long updateUser(User user);
 
-    int deleteUser(int id);
+    long deleteUser(long id);
 
     List<User> getAllUsers();
 }
